Fix integer overflow on 32-bit systems when creating 4GB+ images (#8)
* fix: integer overflow on 32-bit systems when creating 4GB+ images

- Use int64 for size calculations to prevent overflow on 32-bit systems
- Fix file size comparison to use int64 to handle files >2GB

This fixes issue #6 where creating disk images larger than 2GB would fail
on i386 and armv7 systems with "requested size is smaller than minimum
allowed FAT32" error due to integer overflow.



* test: add multi-architecture test infrastructure

- Add test file generation script for various file sizes
- Add Docker-based multi-architecture test suite
- Tests linux/amd64, linux/arm64, linux/386, and linux/arm/v7
- Demonstrates 32-bit fix working correctly

The test suite helps ensure cross-platform compatibility and can catch
architecture-specific issues like integer overflows.



* build: add 32-bit architectures to goreleaser config

- Add linux/386 and linux/arm/v7 to build targets
- Enables building binaries for 32-bit systems
- Cleanup commented code and hooks

This ensures we can distribute binaries that work on 32-bit systems.



* chore: update Go module dependencies

- Update various dependencies to their latest versions
- Remove duplicate/outdated dependency entries
- Updates include security and performance improvements

These dependency updates were used during all testing of the 32-bit fixes.



---------

+++ create.go
@@ -228,7 +228,8 @@ const BlockSize = 512
 const PartitionStart = 2048
 
 func (c *CreateCommand) createDiskImage(tempFileName string) error {
-	espSize := c.partitionMB * MB
+	// Use int64 to avoid overflow on 32-bit systems
+	espSize := int64(c.partitionMB) * MB
 	diskSize := espSize + 4*MB
 	partitionSectors := espSize / BlockSize
 	//partitionEnd := partitionSectors - PartitionStart + 1
@@ -244,7 +245,7 @@ func (c *CreateCommand) createDiskImage(tempFileName string) error {
 		Partitions: []*mbr.Partition{
 			{
 				Start:    PartitionStart,
-				Size:     uint32(partitionSectors),
+				Size:     uint32(partitionSectors), // Note: limits partition to ~2TB
 				Type:     mbr.EFISystem,
 				Bootable: true,
 			},
@@ -342,7 +343,7 @@ func copyFile(src string, dst string, fs filesystem.FileSystem) error {
 	if err != nil {
 		return err
 	}
-	if n != int(info.Size()) {
+	if int64(n) != info.Size() {
 		return fmt.Errorf("error writing output file '%s': %d bytes written, s/b %d\n", dst, n, info.Size())
 	}
 	err = rw.Close()
